requirements.txt
azure-cognitiveservices-speech
openai
python-dotenv
pdfplumber
tkinterdnd2
